Migrated StatusCodeRangeSpecificnessComparer and ResponseTypeInfoCollection tests.
