+++ src/ReqRest.Tests/Internal/StatusCodeRangeExtensionsTests.cs
@@ -1,9 +1,5 @@
 ﻿namespace ReqRest.Tests.Internal
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Threading.Tasks;
-    using System.Linq;
     using Xunit;
     using ReqRest.Http;
     using ReqRest.Tests.Internal.TestData;
